Split rendering code out of Mesh
